[Feature] HTTP Compression (#20)
HTTP Compression and Decompression Implementation. (br, gzip and deflate).

+++ core/src/main/java/io/netty/handler/codec/http2/InboundHttp2ToHttpObjectAdapterBuilder.java
@@ -34,7 +34,7 @@ public final class InboundHttp2ToHttpObjectAdapterBuilder extends AbstractInboun
     }
 
     @Override
-    public InboundHttp2ToHttpObjectAdapterBuilder maxContentLength(int maxContentLength) {
+    public InboundHttp2ToHttpObjectAdapterBuilder maxContentLength(long maxContentLength) {
         return super.maxContentLength(maxContentLength);
     }
 
@@ -55,7 +55,7 @@ public final class InboundHttp2ToHttpObjectAdapterBuilder extends AbstractInboun
 
     @Override
     protected InboundHttp2ToHttpObjectAdapter build(Http2Connection connection,
-                                                    int maxContentLength,
+                                                    long maxContentLength,
                                                     boolean validateHttpHeaders,
                                                     boolean propagateSettings) throws Exception {
